code/requirements.txt
# NeuroGrip Pro V2 — 依赖

# 核心依赖（所有环境）
numpy>=1.21.0
scipy>=1.7.0
pyyaml>=6.0

# 训练环境
# mindspore>=2.7.1    # NPU/GPU 上安装, 参考: https://www.mindspore.cn/install

# 运行环境 (Orange Pi)
# mindspore-lite       # 参考 MindSpore Lite 官方文档安装
pyserial>=3.5         # 串口通信 (臂环)
# smbus2>=0.4.0       # I2C 通信 (PCA9685, 仅 Linux)
